requirements.txt
frappe>=14.0.0
requests>=2.31.0
python-dotenv>=1.0.0
facebook-business>=16.0.0 